Common: BMP: Fix the check issue caused by the extra 2 bytes in the BMP file
Microsoft MU introduced the BMP check function six months ago. This
function performs a series of detections on BMP files. Currently, we
have found that most of our BMP files fail to pass this check.
Therefore, we truncate the last 2 bytes of each problematic BMP file
(this is the root cause of the issue: the actual file size is 2 bytes
larger than the calculated size) to resolve the problem. After testing,
each BMP file can still be opened normally.

Signed-off-by: XiaoYeZi <qaz6750@outlook.com>